refactor : lotto 관련 객체들 패키지 분리
