+++ src/main/java/autolotto/lotto/Lotto.java
@@ -1,5 +1,6 @@
-package autolotto;
+package autolotto.lotto;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
@@ -7,7 +8,7 @@ public class Lotto {
     private final List<Integer> numbers;
 
     public Lotto(List<Integer> numbers) {
-        this.numbers = numbers;
+        this.numbers = new ArrayList<>(numbers);
     }
 
     public int matchCount(List<Integer> comparisonTarget) {
@@ -22,4 +23,7 @@ public class Lotto {
         return this.numbers.size();
     }
 
+    public List<Integer> lottoNumbers() {
+        return new ArrayList<>(this.numbers);
+    }
 }
