+++ src/test/java/autolotto/lotto/LottoTest.java
@@ -1,4 +1,4 @@
-package autolotto;
+package autolotto.lotto;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
